Split /quest/[id] in components
